Create Code adder for membership
This will be continued to be updated as I keep going. This is a "plug and play" kind of approach so you can take snippets that I threw in here to add into different files alternatively you can tell me what you want, and Ill add it and tweak it so it doesn't break.